* commit '13427c01ee4c0f85b8baabb2294fb527013acfc2':
  Do not merge: cherry-pick incorect clip CL from master

+++ WebKit/android/plugins/ANPOpenGLInterface.cpp
@@ -67,8 +67,8 @@ static ANPTextureInfo anp_lockTexture(NPP instance) {
 
     ANPTextureInfo anpInfo;
     anpInfo.textureId = info->m_textureId;
-    anpInfo.width = info->m_width;
-    anpInfo.height = info->m_height;
+    anpInfo.width = (int32_t) info->m_width;
+    anpInfo.height = (int32_t) info->m_height;
     anpInfo.internalFormat = info->m_internalFormat;
     return anpInfo;
 }
